asked the ai to refactor REwww a bit: make some smaller files
