+++ REwww/templates/_projects_pane.html
@@ -1,0 +1,118 @@
+<!-- ══ PROJECTS TAB ══════════════════════════════════════════════════════ -->
+<div class="tab-pane fade show active" id="tab-projects" role="tabpanel">
+
+  <!-- View toolbar -->
+  <div class="tab-toolbar" id="projects-view-toolbar">
+    <button class="btn btn-sm btn-outline-secondary" id="btn-edit-projects">
+      <i class="fas fa-edit"></i> Edit
+    </button>
+    <button class="btn btn-sm btn-outline-secondary" id="btn-refresh-projects"
+            title="Re-scan project directories for available files">
+      <i class="fas fa-sync-alt"></i> Refresh
+    </button>
+  </div>
+
+  <!-- View: run-mode table -->
+  <div id="projects-view">
+    <div class="table-responsive">
+      <table class="table table-sm table-hover projects-table">
+        <thead class="thead-light">
+          <tr>
+            <th>Project</th>
+            <th style="width:14%">Recon</th>
+            <th style="width:12%">MEL</th>
+            <th style="width:12%">Fuzzy</th>
+            <th style="width:26%">Upstream</th>
+            <th style="width:7%">Action</th>
+          </tr>
+        </thead>
+        <tbody>
+          {% for name, info in project_data.items() %}
+          <tr id="proj-row-{{ name }}" data-project-path="{{ info.path }}">
+            <td class="project-name align-middle">{{ name }}</td>
+            <td>
+              <select class="form-control form-control-sm"
+                      id="sel-recon-{{ name }}">
+                {% if not info.recons %}
+                  <option value="">— none found —</option>
+                {% else %}
+                  {% for r in info.recons %}
+                    <option value="{{ r }}">{{ r }}</option>
+                  {% endfor %}
+                {% endif %}
+              </select>
+            </td>
+            <td>
+              <select class="form-control form-control-sm"
+                      id="sel-mel-{{ name }}">
+                {% if not info.mels %}
+                  <option value="">— none found —</option>
+                {% else %}
+                  <option value="">N/A</option>
+                  {% for m in info.mels %}
+                    <option value="{{ m }}">{{ m }}</option>
+                  {% endfor %}
+                {% endif %}
+              </select>
+            </td>
+            <td>
+              <select class="form-control form-control-sm"
+                      id="sel-fuzzy-{{ name }}">
+                {% if not info.fuzzies %}
+                  <option value="">— none found —</option>
+                {% else %}
+                  <option value="">N/A</option>
+                  {% for f in info.fuzzies %}
+                    <option value="{{ f }}">{{ f }}</option>
+                  {% endfor %}
+                {% endif %}
+              </select>
+            </td>
+            <td>
+              <input type="text" class="form-control form-control-sm"
+                     id="inp-upstream-{{ name }}"
+                     placeholder="e.g. PIWR: PWR, it; PWR: PIR, fr">
+            </td>
+            <td class="align-middle">
+              <button class="btn btn-primary btn-sm run-btn"
+                      data-project="{{ name }}">
+                <i class="fas fa-play"></i> Run
+              </button>
+            </td>
+          </tr>
+          {% endfor %}
+        </tbody>
+      </table>
+    </div>
+  </div><!-- /projects-view -->
+
+  <!-- Edit toolbar + table (hidden until Edit clicked) -->
+  <div id="projects-edit" style="display:none">
+    <div class="tab-toolbar">
+      <button class="btn btn-sm btn-success" id="btn-save-projects">
+        <i class="fas fa-save"></i> Save
+      </button>
+      <button class="btn btn-sm btn-secondary" id="btn-cancel-projects">
+        Cancel
+      </button>
+      <button class="btn btn-sm btn-outline-primary" id="btn-add-project">
+        + Add project
+      </button>
+    </div>
+    <div class="table-responsive">
+      <table class="table table-sm table-bordered projects-table" id="projects-edit-table">
+        <thead class="thead-light">
+          <tr>
+            <th class="actions-col" style="width:1%;white-space:nowrap"></th>
+            <th>Project name</th>
+            <th>Directory path</th>
+          </tr>
+        </thead>
+        <tbody id="projects-edit-body">
+          <!-- populated by setupProjectsEdit() -->
+        </tbody>
+      </table>
+    </div>
+  </div><!-- /projects-edit -->
+
+</div><!-- /tab-projects -->
