+++ REwww/templates/about.html
@@ -57,7 +57,7 @@
 </p>
 <p>
   The current implementation is written in Python, the core functionality recreated by
-  Zhang, subsequently amplified and embellished by Zhang and Lowe.
+  Zhang, subsequently amplified and embellished by Lowe, Mazaudon, and Zhang.
 </p>
 <p>
   This web interface was written by Lowe and uses the
